Added logging to methods with implementation

+++ src/main/java/by/tms/featuric/service/impl/AnswerServiceImpl.java
@@ -1,13 +1,18 @@
 package by.tms.featuric.service.impl;
 
 import by.tms.featuric.entity.FtrcAnswer;
+import by.tms.featuric.exception.ExistsException;
+import by.tms.featuric.exception.IndexOutOfBoundsException;
+import by.tms.featuric.exception.NotFoundException;
 import by.tms.featuric.repository.FtrcAnswerRepository;
 import by.tms.featuric.service.interfaces.AnswerService;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.math.BigInteger;
 
+@Slf4j
 @Service
 public class AnswerServiceImpl implements AnswerService {
 
@@ -21,11 +26,16 @@ public class AnswerServiceImpl implements AnswerService {
 
     @Override
     public FtrcAnswer save(FtrcAnswer ftrcAnswer) {
+        log.info("Method - save answer");
+        if(ftrcAnswerRepository.findByDescription(ftrcAnswer.getDescription()) != null){
+            throw new ExistsException("Answer exists");
+        }
         return ftrcAnswerRepository.save(ftrcAnswer);
     }
 
     @Override
     public FtrcAnswer update(BigInteger id, FtrcAnswer ftrcAnswerRequest) {
+        log.info("Method - update answer");
         ftrcAnswerRepository.findById(id).map(ftrcAnswer -> {
             ftrcAnswer.setDescription(ftrcAnswerRequest.getDescription());
             ftrcAnswer.setRightAnswer(ftrcAnswerRequest.isRightAnswer());
@@ -37,11 +47,17 @@ public class AnswerServiceImpl implements AnswerService {
 
     @Override
     public void deleteAnswerById(BigInteger id) {
-        ftrcAnswerRepository.deleteById(id);
+        log.info("Method - delete answer by ID");
+        if (ftrcAnswerRepository.findById(id).isPresent()) {
+            ftrcAnswerRepository.deleteById(id);
+        } else {
+            throw new NotFoundException("Answer with id " + id + " not found");
+        }
     }
 
     @Override
     public FtrcAnswer findAnswerById(BigInteger id) {
-        return ftrcAnswerRepository.findById(id).orElseThrow(ArrayIndexOutOfBoundsException::new);
+        log.info("Method - find answer by ID");
+        return ftrcAnswerRepository.findById(id).orElseThrow(IndexOutOfBoundsException::new);
     }
 }
