+++ src/main/java/by/tms/featuric/service/impl/UserServiceImpl.java
@@ -1,18 +1,22 @@
 package by.tms.featuric.service.impl;
 
 import by.tms.featuric.entity.FtrcUser;
+import by.tms.featuric.exception.ExistsException;
+import by.tms.featuric.exception.IndexOutOfBoundsException;
+import by.tms.featuric.exception.NotFoundException;
 import by.tms.featuric.repository.FtrcUserRepository;
 import by.tms.featuric.service.interfaces.UserService;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.math.BigInteger;
 import java.util.List;
 
+@Slf4j
 @Service
 public class UserServiceImpl implements UserService, UserDetailsService {
 
@@ -27,12 +31,17 @@ public class UserServiceImpl implements UserService, UserDetailsService {
 
     @Override
     public FtrcUser save(FtrcUser ftrcUser) {
+        log.info("Method - save user");
+        if (ftrcUserRepository.findFtrcUserByLogin(ftrcUser.getLogin()) != null) {
+            throw new ExistsException("User exists");
+        }
         ftrcUser.setPassword(passwordEncoder.encode(ftrcUser.getPassword()));
         return ftrcUserRepository.save(ftrcUser);
     }
 
     @Override
     public FtrcUser update(BigInteger id, FtrcUser ftrcUserRequest) {
+        log.info("Method - update user");
         ftrcUserRepository.findById(id).map(ftrcUser -> {
             ftrcUser.setFirstName(ftrcUserRequest.getFirstName());
             ftrcUser.setLastName(ftrcUserRequest.getLastName());
@@ -49,35 +58,48 @@ public class UserServiceImpl implements UserService, UserDetailsService {
 
     @Override
     public void deleteById(BigInteger id) {
-        ftrcUserRepository.deleteById(id);
+        log.info("Method - delete user by ID");
+        if (ftrcUserRepository.findById(id).isPresent()) {
+            ftrcUserRepository.deleteById(id);
+        }
+        throw new NotFoundException("User with id " + id + " not found");
     }
 
     @Override
     public List<FtrcUser> getAllUsers() {
+        log.info("Method - get all users");
         return ftrcUserRepository.findAll();
     }
 
     @Override
     public FtrcUser findUserById(BigInteger id) {
-        return ftrcUserRepository.findById(id).orElseThrow(ArrayIndexOutOfBoundsException::new);
+        log.info("Method - find user by ID");
+        return ftrcUserRepository.findById(id).orElseThrow(IndexOutOfBoundsException::new);
     }
 
     @Override
     public FtrcUser findUserByFirstNameAndLastName(String firstName, String lastName) {
+        log.info("Method - find user by first name and last name");
         return ftrcUserRepository.findFtrcUserByFirstNameAndLastName(firstName, lastName);
     }
 
     @Override
     public FtrcUser findUserByLogin(String login) {
-        return ftrcUserRepository.findFtrcUserByLogin(login);
+        log.info("Method - find user by login");
+        FtrcUser user = ftrcUserRepository.findFtrcUserByLogin(login);
+        if (user == null) {
+            throw new NotFoundException("User with login " + login + " not found");
+        }
+        return user;
     }
 
     //UserDetailsService
     @Override
-    public UserDetails loadUserByUsername(String login) throws UsernameNotFoundException {
+    public UserDetails loadUserByUsername(String login) throws NotFoundException {
+        log.info("Method - load user by login(username)//security");
         FtrcUser user = ftrcUserRepository.findFtrcUserByLogin(login);
-        if(user == null){
-            throw new UsernameNotFoundException("User not found");
+        if (user == null) {
+            throw new NotFoundException("User not found");
         }
         return (UserDetails) user;
     }
